+++ Tests/Browser.App.Tests/Src/BrowserAppTests.cs
@@ -186,21 +186,6 @@ public class BrowserAppTests : DisposableBase, IClassFixture<AppServiceFixture>
         Assert.NotNull(browser);
     }
     
-    [Fact]
-    public async Task RemovePage_LastPage_MainWindowShouldBeClosed()
-    {
-        // Arrange
-        var browser = _appService.GetService<IBrowser>();
-        var mainWindow = _appService.GetService<IMainWindow>();
-        var currentPage = browser.CurrentPage.Value;
-        
-        // Act
-        await browser.RemovePage(currentPage.Id);
-        
-        // Assert
-        mainWindow.Received().Close();
-    }
-    
 
     protected override void Dispose(bool disposing)
     {
